pyproject.toml
[project]
name = "rembg-service"
version = "0.1.0"
description = "A Flask service for AI-based image background removal"
requires-python = ">=3.11"
dependencies = [
    "flask==3.0.3",
    "flask-cors==5.0.0",
    "rembg==2.0.57",
    "pillow==10.4.0",
]

[tool.uv]
dev-dependencies = []